requirements.txt
awsiotsdk
pygame
awscert
sqlite